- Refactored the code (BRAINSTORM-2694) - Fixed groupareas=0 bug (BRAINSTORM-2716)
