+++ test/base/input/area_with_radius_yes_groupareas.get
@@ -1,0 +1,1 @@
+GET /timeseries?starttime=200808051200&area=Helsinki:10&precision=double&param=name,time,lon,lat,Temperature&groupareas=0 HTTP/1.0
