+++ test/base/input/observation_place_radius20.get
@@ -1,0 +1,1 @@
+GET /timeseries?format=ascii&places=Tampere:20&param=place,name,time,Temperature&timestep=60&producer=opendata&starttime=201308050800&endtime=201308051100 HTTP/1.0
