+++ test/base/input/place_kumpula_radius20.get
@@ -1,0 +1,1 @@
+GET /timeseries?starttime=200808051200&place=Kumpula:20&precision=double&param=name,time,Temperature&producer=pal HTTP/1.0
