+++ test/base/output/observation_place_radius20.get
@@ -1,0 +1,4 @@
+[101118 101124 101311 151049] Tampere 20130805T080000 [17 17 17 17]
+[101118 101124 101311 151049] Tampere 20130805T090000 [17 18 17 18]
+[101118 101124 101311 151049] Tampere 20130805T100000 [19 20 18 20]
+[101118 101124 101311 151049] Tampere 20130805T110000 [21 21 19 20]
