+++ test/base/output/place_kumpula_radius20.get
@@ -1,0 +1,25 @@
+Kumpula 20080805T120000 [14.8 15.1 14.6 14.8 15.3 15.1 15.5]
+Kumpula 20080805T130000 [14.4 14.7 14.3 14.4 14.9 14.9 15.2]
+Kumpula 20080805T140000 [14.4 14.8 14.4 14.6 15.1 15.1 15.5]
+Kumpula 20080805T150000 [14.4 14.7 14.4 14.6 15.1 15.2 15.6]
+Kumpula 20080805T160000 [14.5 14.9 14.5 14.6 15.1 15.1 15.4]
+Kumpula 20080805T170000 [14.8 15.2 14.7 14.9 15.3 15.2 15.4]
+Kumpula 20080805T180000 [15.1 15.4 14.9 15.1 15.4 15.2 15.4]
+Kumpula 20080805T190000 [15.3 15.8 14.9 15.1 15.6 15.0 15.2]
+Kumpula 20080805T200000 [15.5 16.0 14.8 15.2 15.8 14.7 15.0]
+Kumpula 20080805T210000 [14.6 15.2 13.6 14.1 14.8 13.3 13.8]
+Kumpula 20080805T220000 [14.7 15.4 13.4 14.0 14.7 13.0 13.5]
+Kumpula 20080805T230000 [14.8 15.5 13.2 13.8 14.6 12.7 13.2]
+Kumpula 20080806T000000 [14.9 15.6 13.0 13.7 14.5 12.3 12.9]
+Kumpula 20080806T010000 [14.8 15.7 12.8 13.5 14.4 12.0 12.6]
+Kumpula 20080806T020000 [14.8 15.7 12.6 13.4 14.3 11.7 12.3]
+Kumpula 20080806T030000 [14.7 15.7 12.3 13.2 14.2 11.5 12.0]
+Kumpula 20080806T040000 [14.3 15.2 11.9 12.7 13.7 11.1 11.6]
+Kumpula 20080806T050000 [13.8 14.8 11.4 12.3 13.2 10.7 11.2]
+Kumpula 20080806T060000 [13.3 14.3 10.9 11.8 12.8 10.3 10.8]
+Kumpula 20080806T070000 [13.7 14.5 12.0 12.6 13.3 11.5 11.9]
+Kumpula 20080806T080000 [14.2 14.6 13.1 13.4 13.8 12.7 13.0]
+Kumpula 20080806T090000 [14.6 14.7 14.1 14.2 14.3 13.9 14.1]
+Kumpula 20080806T100000 [15.5 15.3 15.4 15.4 15.2 15.3 15.3]
+Kumpula 20080806T110000 [16.5 16.1 16.8 16.6 16.3 16.7 16.7]
+Kumpula 20080806T120000 [17.5 17.0 18.1 17.8 17.4 18.0 17.9]
