+++ timeseries/Query.cpp
@@ -5,29 +5,13 @@
 // ======================================================================
 
 #include "Query.h"
-#include "Config.h"
-#include "Hash.h"
-#include "State.h"
-#include <engines/geonames/Engine.h>
-#ifndef WITHOUT_OBSERVATION
-#include <engines/observation/Engine.h>
-#endif
-#include <boost/algorithm/string/classification.hpp>
-#include <boost/algorithm/string/split.hpp>
-#include <boost/foreach.hpp>
-#include <gis/CoordinateTransformation.h>
-#include <grid-files/common/GeneralFunctions.h>
-#include <grid-files/common/ShowFunction.h>
+#include <spine/Convenience.h>
 #include <macgyver/AnsiEscapeCodes.h>
 #include <macgyver/DistanceParser.h>
-#include <macgyver/Exception.h>
-#include <macgyver/StringConversion.h>
-#include <macgyver/TimeParser.h>
-#include <newbase/NFmiPoint.h>
-#include <spine/Convenience.h>
-#include <timeseries/ParameterFactory.h>
-#include <timeseries/TimeSeriesGeneratorOptions.h>
-#include <algorithm>
+#include <gis/CoordinateTransformation.h>
+#include <engines/observation/ExternalAndMobileProducerId.h>
+#include <engines/querydata/Engine.h>
+#include "Config.h"
 
 using namespace std;
 
@@ -41,6 +25,65 @@ const char* default_timezone = "localtime";
 
 namespace
 {
+
+void set_max_agg_interval_behind(TS::DataFunction& func, unsigned int& max_interval)
+{
+  try
+  {
+	if (func.type() == TS::FunctionType::TimeFunction)
+      {
+        if (max_interval < func.getAggregationIntervalBehind())
+          max_interval = func.getAggregationIntervalBehind();
+	  }
+  }
+  catch (...)
+  {
+    throw Fmi::Exception::Trace(BCP, "Operation failed!");
+  }
+}
+
+void set_max_agg_interval_ahead(TS::DataFunction& func, unsigned int& max_interval)
+{
+  try
+  {
+	if (func.type() == TS::FunctionType::TimeFunction)
+      {
+        if (max_interval < func.getAggregationIntervalAhead())
+          max_interval = func.getAggregationIntervalAhead();
+	  }
+  }
+  catch (...)
+  {
+    throw Fmi::Exception::Trace(BCP, "Operation failed!");
+  }
+}
+
+void set_agg_interval_behind(TS::DataFunction& func, unsigned int interval)
+{
+  try
+  {
+	if (func.getAggregationIntervalBehind() == std::numeric_limits<unsigned int>::max())
+	  func.setAggregationIntervalBehind(interval);
+  }
+  catch (...)
+  {
+    throw Fmi::Exception::Trace(BCP, "Operation failed!");
+  }
+}
+
+void set_agg_interval_ahead(TS::DataFunction& func, unsigned int interval)
+{
+  try
+  {
+	if (func.getAggregationIntervalAhead() == std::numeric_limits<unsigned int>::max())
+	  func.setAggregationIntervalAhead(interval);
+  }
+  catch (...)
+  {
+    throw Fmi::Exception::Trace(BCP, "Operation failed!");
+  }
+}
+
 void add_sql_data_filter(const Spine::HTTP::Request& req,
                          const std::string& param_name,
                          TS::DataFilter& dataFilter)
@@ -79,6 +122,7 @@ Fmi::ValueFormatterParam valueformatter_params(const Spine::HTTP::Request& req)
 
 }  // namespace
 
+
 // ----------------------------------------------------------------------
 /*!
  * \brief The constructor parses the query string
@@ -86,7 +130,7 @@ Fmi::ValueFormatterParam valueformatter_params(const Spine::HTTP::Request& req)
 // ----------------------------------------------------------------------
 
 Query::Query(const State& state, const Spine::HTTP::Request& req, Config& config)
-    : valueformatter(valueformatter_params(req)), timeAggregationRequested(false)
+  : ObsQueryParams(req), valueformatter(valueformatter_params(req)), timeAggregationRequested(false)
 {
   try
   {
@@ -109,79 +153,9 @@ Query::Query(const State& state, const Spine::HTTP::Request& req, Config& config
 
     forecastSource = Spine::optional_string(req.getParameter("source"), "");
 
-    // ### Attribute list ( attr=name1:value1,name2:value2,name3:$(alias1) )
+	parse_attr(req);
 
-    string attr = Spine::optional_string(req.getParameter("attr"), "");
-    if (!attr.empty())
-    {
-      bool ind = true;
-      uint loopCount = 0;
-      while (ind)
-      {
-        loopCount++;
-        if (loopCount > 10)
-          throw Fmi::Exception(BCP, "The alias definitions seem to contain an eternal loop!");
-
-        ind = false;
-        std::string alias;
-        if (itsAliasFileCollectionPtr->replaceAlias(attr, alias))
-        {
-          attr = alias;
-          ind = true;
-        }
-      }
-
-      std::vector<std::string> partList;
-      splitString(attr, ';', partList);
-      for (const auto& part : partList)
-      {
-        std::vector<std::string> list;
-        splitString(part, ':', list);
-        if (list.size() == 2)
-        {
-          std::string name = list[0];
-          std::string value = list[1];
-
-          attributeList.addAttribute(name, value);
-        }
-      }
-    }
-
-    // attributeList.print(std::cout,0,0);
-
-    T::Attribute* v1 = attributeList.getAttributeByNameEnd("Grib1.IndicatorSection.EditionNumber");
-    T::Attribute* v2 = attributeList.getAttributeByNameEnd("Grib2.IndicatorSection.EditionNumber");
-
-    T::Attribute* lat = attributeList.getAttributeByNameEnd("LatitudeOfFirstGridPoint");
-    T::Attribute* lon = attributeList.getAttributeByNameEnd("LongitudeOfFirstGridPoint");
-
-    if (v1 != nullptr && lat != nullptr && lon != nullptr)
-    {
-      // Using coordinate that is inside the GRIB1 grid
-
-      double latitude = toDouble(lat->mValue.c_str()) / 1000;
-      double longitude = toDouble(lon->mValue.c_str()) / 1000;
-
-      std::string val = std::to_string(latitude) + "," + std::to_string(longitude);
-      Spine::HTTP::Request tmpReq;
-      tmpReq.addParameter("latlon", val);
-      loptions.reset(
-          new Engine::Geonames::LocationOptions(state.getGeoEngine().parseLocations(tmpReq)));
-    }
-    else if (v2 != nullptr && lat != nullptr && lon != nullptr)
-    {
-      // Using coordinate that is inside the GRIB2 grid
-
-      double latitude = toDouble(lat->mValue.c_str()) / 1000000;
-      double longitude = toDouble(lon->mValue.c_str()) / 1000000;
-
-      std::string val = std::to_string(latitude) + "," + std::to_string(longitude);
-      Spine::HTTP::Request tmpReq;
-      tmpReq.addParameter("latlon", val);
-      loptions.reset(
-          new Engine::Geonames::LocationOptions(state.getGeoEngine().parseLocations(tmpReq)));
-    }
-    else
+	if(!parse_grib_loptions(state))
     {
       loptions.reset(
           new Engine::Geonames::LocationOptions(state.getGeoEngine().parseLocations(req)));
@@ -221,11 +195,6 @@ Query::Query(const State& state, const Spine::HTTP::Request& req, Config& config
     std::string endtime = Spine::optional_string(req.getParameter("endtime"), "");
     endtimeOptionGiven = (endtime != "now");
 
-#ifndef WITHOUT_OBSERVATION
-    latestObservation = (endtime == "now");
-    allplaces = (Spine::optional_string(req.getParameter("places"), "") == "all");
-#endif
-
     debug = Spine::optional_bool(req.getParameter("debug"), false);
 
     timezone = Spine::optional_string(req.getParameter("tz"), default_timezone);
@@ -253,44 +222,14 @@ Query::Query(const State& state, const Spine::HTTP::Request& req, Config& config
         Spine::optional_string(req.getParameter("maxdistance"), config.defaultMaxDistance());
 
     keyword = Spine::optional_string(req.getParameter("keyword"), "");
-    auto searchName = req.getParameterList("inkeyword");
-    if (!searchName.empty())
-    {
-      for (const std::string& key : searchName)
-      {
-        Locus::QueryOptions opts;
-        opts.SetLanguage(language);
-        Spine::LocationList places = state.getGeoEngine().keywordSearch(opts, key);
-        if (places.empty())
-          throw Fmi::Exception(BCP, "No locations for keyword " + std::string(key) + " found");
-        inKeywordLocations.insert(inKeywordLocations.end(), places.begin(), places.end());
-      }
-    }
+
+	parse_inkeyword_locations(req, state);
 
     findnearestvalidpoint = Spine::optional_bool(req.getParameter("findnearestvalid"), false);
 
-    boost::optional<std::string> tmp = req.getParameter("origintime");
-    if (tmp)
-    {
-      if (*tmp == "latest" || *tmp == "newest")
-        origintime = boost::posix_time::ptime(boost::date_time::pos_infin);
-      else if (*tmp == "oldest")
-        origintime = boost::posix_time::ptime(boost::date_time::neg_infin);
-      else
-        origintime = Fmi::TimeParser::parse(*tmp);
-    }
-
-#ifndef WITHOUT_OBSERVARION
-    Query::parse_producers(req, state.getQEngine(), state.getGridEngine(), state.getObsEngine());
-#else
-    Query::parse_producers(req, state.getQEngine(), state.getGridEngine());
-#endif
-
-#ifndef WITHOUT_OBSERVATION
-    Query::parse_parameters(req, state.getObsEngine());
-#else
-    Query::parse_parameters(req);
-#endif
+	parse_origintime(req);
+    parse_producers(req, state);
+    parse_parameters(req);
 
     Query::parse_levels(req);
 
@@ -317,118 +256,6 @@ Query::Query(const State& state, const Spine::HTTP::Request& req, Config& config
 
     timeformatter.reset(Fmi::TimeFormatter::create(timeformat));
 
-#ifndef WITHOUT_OBSERVATION
-    // observation params
-    numberofstations = Spine::optional_int(req.getParameter("numberofstations"), 1);
-    std::string opt_stationgroups = Spine::optional_string(req.getParameter("stationgroups"), "");
-    if (!opt_stationgroups.empty())
-      boost::algorithm::split(stationgroups, opt_stationgroups, boost::algorithm::is_any_of(","));
-#endif
-
-#ifndef WITHOUT_OBSERVATION
-    auto name = req.getParameter("fmisid");
-    if (name)
-    {
-      const string fmisidreq = *name;
-      vector<string> parts;
-      boost::algorithm::split(parts, fmisidreq, boost::algorithm::is_any_of(","));
-
-      for (const string& sfmisid : parts)
-      {
-        int f = Fmi::stoi(sfmisid);
-        this->fmisids.push_back(f);
-      }
-    }
-
-    // sid is an alias for fmisid
-    name = req.getParameter("sid");
-    if (name)
-    {
-      const string sidreq = *name;
-      vector<string> parts;
-      boost::algorithm::split(parts, sidreq, boost::algorithm::is_any_of(","));
-
-      for (const string& ssid : parts)
-      {
-        int f = Fmi::stoi(ssid);
-        this->fmisids.push_back(f);
-      }
-    }
-#endif
-
-#ifndef WITHOUT_OBSERVATION
-    name = req.getParameter("wmo");
-    if (name)
-    {
-      const string wmoreq = *name;
-      vector<string> parts;
-      boost::algorithm::split(parts, wmoreq, boost::algorithm::is_any_of(","));
-
-      for (const string& swmo : parts)
-      {
-        int w = Fmi::stoi(swmo);
-        this->wmos.push_back(w);
-      }
-    }
-#endif
-
-#ifndef WITHOUT_OBSERVATION
-    name = req.getParameter("lpnn");
-    if (name)
-    {
-      const string lpnnreq = *name;
-      vector<string> parts;
-      boost::algorithm::split(parts, lpnnreq, boost::algorithm::is_any_of(","));
-
-      for (const string& slpnn : parts)
-      {
-        int l = Fmi::stoi(slpnn);
-        this->lpnns.push_back(l);
-      }
-    }
-#endif
-
-#ifndef WITHOUT_OBSERVATION
-    if (!!req.getParameter("bbox"))
-    {
-      string bbox = *req.getParameter("bbox");
-      vector<string> parts;
-      boost::algorithm::split(parts, bbox, boost::algorithm::is_any_of(","));
-      std::string lat2(parts[3]);
-      auto radius_pos = lat2.find(':');
-      if (radius_pos != string::npos)
-        lat2.resize(radius_pos);
-
-      // Bounding box must contain exactly 4 elements
-      if (parts.size() != 4)
-      {
-        throw Fmi::Exception(BCP, "Invalid bounding box '" + bbox + "'!");
-      }
-
-      if (!parts[0].empty())
-        boundingBox["minx"] = Fmi::stod(parts[0]);
-
-      if (!parts[1].empty())
-        boundingBox["miny"] = Fmi::stod(parts[1]);
-
-      if (!parts[2].empty())
-        boundingBox["maxx"] = Fmi::stod(parts[2]);
-
-      if (!parts[3].empty())
-        boundingBox["maxy"] = Fmi::stod(lat2);
-    }
-    // Data filter options
-    add_sql_data_filter(req, "station_id", dataFilter);
-    add_sql_data_filter(req, "data_quality", dataFilter);
-    // Sounding filtering options
-    add_sql_data_filter(req, "sounding_type", dataFilter);
-    add_sql_data_filter(req, "significance", dataFilter);
-
-    std::string dataCache = Spine::optional_string(req.getParameter("useDataCache"), "true");
-    useDataCache = (dataCache == "true" || dataCache == "1");
-
-#endif
-
     if (!!req.getParameter("weekday"))
     {
       const string sweekdays = *req.getParameter("weekday");
@@ -454,97 +281,75 @@ Query::Query(const State& state, const Spine::HTTP::Request& req, Config& config
   }
 }
 
-// ----------------------------------------------------------------------
-/*!
- * \brief Parse producer/model options
- */
-// ----------------------------------------------------------------------
-
-#ifndef WITHOUT_OBSERVATION
-void Query::parse_producers(const Spine::HTTP::Request& theReq,
-                            const Engine::Querydata::Engine& theQEngine,
-                            const Engine::Grid::Engine* theGridEngine,
-                            const Engine::Observation::Engine* theObsEngine)
-#else
-void Query::parse_producers(const Spine::HTTP::Request& theReq,
-                            const Engine::Querydata::Engine& theQEngine,
-                            const Engine::Grid::Engine& theGridEngine)
-#endif
+void Query::parse_producers(const Spine::HTTP::Request& theReq, const State& theState)
 {
   try
-  {
-    string opt = Spine::optional_string(theReq.getParameter("model"), "");
+	{
+	  string opt = Spine::optional_string(theReq.getParameter("model"), "");
+	  string opt2 = Spine::optional_string(theReq.getParameter("producer"), "");
+	  
+	  if (opt == "default" || opt2 == "default")
+		{
+		  // Use default if it's forced by any option
+		  return;
+		}
+	  
+	  // observation uses stationtype-parameter
+	  if (opt.empty() && opt2.empty())
+		opt2 = Spine::optional_string(theReq.getParameter("stationtype"), "");
+	  
+	  std::list<std::string> resultProducers;
+	  
+	  // Handle time separation:: either 'model' or 'producer' keyword used
+	  if (!opt.empty())
+		boost::algorithm::split(resultProducers, opt, boost::algorithm::is_any_of(";"));
+	  else if (!opt2.empty())
+		boost::algorithm::split(resultProducers, opt2, boost::algorithm::is_any_of(";"));
+	  
+	  for (auto& p : resultProducers)
+		{
+		  boost::algorithm::to_lower(p);
+		  if (p == "itmf")
+			{
+			  p = Engine::Observation::FMI_IOT_PRODUCER;
+			  iot_producer_specifier = "itmf";
+			}
+		}
+	  
+	  // Verify the producer names are valid
+	  const auto obsProducers = getObsProducers(theState);
+	  
+	  for (const auto& p : resultProducers)
+		{
+		  const auto& qEngine = theState.getQEngine();
+		  bool ok = qEngine.hasProducer(p);
+		  if (!obsProducers.empty() && !ok)
+			ok = (obsProducers.find(p) != obsProducers.end());
+		  const auto* gridEngine = theState.getGridEngine();
+		  if (!ok && gridEngine)
+			ok = gridEngine->isGridProducer(p);
+		  
+		  if (!ok)
+			throw Fmi::Exception(BCP, "Unknown producer name '" + p + "'");
+		}
 
-    string opt2 = Spine::optional_string(theReq.getParameter("producer"), "");
-
-    if (opt == "default" || opt2 == "default")
-    {
-      // Use default if it's forced by any option
-      return;
-    }
-
-    // observation uses stationtype-parameter
-    if (opt.empty() && opt2.empty())
-      opt2 = Spine::optional_string(theReq.getParameter("stationtype"), "");
-
-    std::list<std::string> resultProducers;
-
-    // Handle time separation:: either 'model' or 'producer' keyword used
-    if (!opt.empty())
-      boost::algorithm::split(resultProducers, opt, boost::algorithm::is_any_of(";"));
-    else if (!opt2.empty())
-      boost::algorithm::split(resultProducers, opt2, boost::algorithm::is_any_of(";"));
-
-    for (auto& p : resultProducers)
-    {
-      boost::algorithm::to_lower(p);
-      if (p == "itmf")
-      {
-        p = Engine::Observation::FMI_IOT_PRODUCER;
-        iot_producer_specifier = "itmf";
-      }
-    }
-
-    // Verify the producer names are valid
-
-#ifndef WITHOUT_OBSERVATION
-    std::set<std::string> observations;
-    if (theObsEngine != nullptr)
-      observations = theObsEngine->getValidStationTypes();
-#endif
-
-    for (const auto& p : resultProducers)
-    {
-      bool ok = theQEngine.hasProducer(p);
-#ifndef WITHOUT_OBSERVATION
-      if (!ok)
-        ok = (observations.find(p) != observations.end());
-#endif
-      if (!ok && theGridEngine)
-        ok = theGridEngine->isGridProducer(p);
-
-      if (!ok)
-        throw Fmi::Exception(BCP, "Unknown producer name '" + p + "'");
-    }
-
-    // Now split into location parts
-
-    for (const auto& tproducers : resultProducers)
-    {
-      AreaProducers producers;
-      boost::algorithm::split(producers, tproducers, boost::algorithm::is_any_of(","));
-
-      // FMI producer is deprecated, use OPENDATA instead
-      // std::replace(producers.begin(), producers.end(), std::string("fmi"),
-      // std::string("opendata"));
-
-      timeproducers.push_back(producers);
-    }
-  }
+	  // Now split into location parts	  
+	  for (const auto& tproducers : resultProducers)
+		{
+		  AreaProducers producers;
+		  boost::algorithm::split(producers, tproducers, boost::algorithm::is_any_of(","));
+		  
+		  // FMI producer is deprecated, use OPENDATA instead
+		  // std::replace(producers.begin(), producers.end(), std::string("fmi"),
+		  // std::string("opendata"));
+		  
+		  timeproducers.push_back(producers);
+		}
+	}
   catch (...)
-  {
-    throw Fmi::Exception::Trace(BCP, "Operation failed!");
-  }
+	{
+	  throw Fmi::Exception::Trace(BCP, "Operation failed!");
+	}
 }
 
 // ----------------------------------------------------------------------
@@ -647,12 +452,63 @@ void Query::parse_precision(const Spine::HTTP::Request& req, const Config& confi
   }
 }
 
-#ifndef WITHOUT_OBSERVATION
-void Query::parse_parameters(const Spine::HTTP::Request& theReq,
-                             const Engine::Observation::Engine* theObsEngine)
-#else
+void Query::parse_aggregation_intervals(const Spine::HTTP::Request& theReq)
+{
+  try
+  {
+    std::string aggregationIntervalStringBehind =
+        Spine::optional_string(theReq.getParameter("interval"), "0m");
+    std::string aggregationIntervalStringAhead = ("0m");
+
+    // check if second aggregation interval is defined
+    auto agg_pos = aggregationIntervalStringBehind.find(':');
+    if (agg_pos != string::npos)
+    {
+      aggregationIntervalStringAhead = aggregationIntervalStringBehind.substr(agg_pos + 1);
+      aggregationIntervalStringBehind.resize(agg_pos);
+    }
+
+    int agg_interval_behind = Spine::duration_string_to_minutes(aggregationIntervalStringBehind);
+    int agg_interval_ahead = Spine::duration_string_to_minutes(aggregationIntervalStringAhead);
+
+    if (agg_interval_behind < 0 || agg_interval_ahead < 0)
+      throw Fmi::Exception(BCP, "The 'interval' option must be positive!");
+
+    // set aggregation interval if it has not been set in parameter parser
+    for (const TS::ParameterAndFunctions& paramfuncs : poptions.parameterFunctions())
+    {
+	  set_agg_interval_behind(const_cast<TS::DataFunction&>(paramfuncs.functions.innerFunction), agg_interval_behind);
+	  set_agg_interval_ahead(const_cast<TS::DataFunction&>(paramfuncs.functions.innerFunction), agg_interval_ahead);
+	  set_agg_interval_behind(const_cast<TS::DataFunction&>(paramfuncs.functions.outerFunction), agg_interval_behind);
+	  set_agg_interval_ahead(const_cast<TS::DataFunction&>(paramfuncs.functions.outerFunction), agg_interval_ahead);
+    }
+
+    // store maximum aggregation intervals per parameter for later use
+    for (const TS::ParameterAndFunctions& paramfuncs : poptions.parameterFunctions())
+    {
+      std::string paramname(paramfuncs.parameter.name());
+
+      if (maxAggregationIntervals.find(paramname) == maxAggregationIntervals.end())
+        maxAggregationIntervals.insert(make_pair(paramname, AggregationInterval(0, 0)));
+
+	  set_max_agg_interval_behind(const_cast<TS::DataFunction&>(paramfuncs.functions.innerFunction), maxAggregationIntervals[paramname].behind);
+	  set_max_agg_interval_ahead(const_cast<TS::DataFunction&>(paramfuncs.functions.innerFunction), maxAggregationIntervals[paramname].ahead);
+	  set_max_agg_interval_behind(const_cast<TS::DataFunction&>(paramfuncs.functions.outerFunction), maxAggregationIntervals[paramname].behind);
+	  set_max_agg_interval_ahead(const_cast<TS::DataFunction&>(paramfuncs.functions.outerFunction), maxAggregationIntervals[paramname].ahead);
+
+      if (maxAggregationIntervals[paramname].behind > 0 ||
+          maxAggregationIntervals[paramname].ahead > 0)
+        timeAggregationRequested = true;
+    }
+  }
+  catch (...)
+  {
+    throw Fmi::Exception::Trace(BCP, "Operation failed!");
+  }
+}
+
+
 void Query::parse_parameters(const Spine::HTTP::Request& theReq)
-#endif
 {
   try
   {
@@ -712,34 +568,6 @@ void Query::parse_parameters(const Spine::HTTP::Request& theReq)
         tmpNames = names;
     }
 
-#ifndef WITHOUT_OBSERVATION
-
-    // Determine whether any producer implies observations are needed
-
-    // TODO: The boolean is unused, is the code buggy?
-
-    bool obsProducersExist = false;
-
-    if (theObsEngine != nullptr)
-    {
-      std::set<std::string> obsEngineStationTypes = theObsEngine->getValidStationTypes();
-      for (const auto& areaproducers : timeproducers)
-      {
-        if (obsProducersExist)
-          break;
-
-        for (const auto& producer : areaproducers)
-        {
-          if (obsEngineStationTypes.find(producer) != obsEngineStationTypes.end())
-          {
-            obsProducersExist = true;
-            break;
-          }
-        }
-      }
-    }
-#endif
-
     // Validate and convert
     for (const string& paramname : names)
     {
@@ -757,82 +585,150 @@ void Query::parse_parameters(const Spine::HTTP::Request& theReq)
     }
     poptions.expandParameter("data_source");
 
-    std::string aggregationIntervalStringBehind =
-        Spine::optional_string(theReq.getParameter("interval"), "0m");
-    std::string aggregationIntervalStringAhead = ("0m");
+	parse_aggregation_intervals(theReq);
 
-    // check if second aggregation interval is defined
-    auto agg_pos = aggregationIntervalStringBehind.find(':');
-    if (agg_pos != string::npos)
-    {
-      aggregationIntervalStringAhead = aggregationIntervalStringBehind.substr(agg_pos + 1);
-      aggregationIntervalStringBehind.resize(agg_pos);
-    }
-
-    int agg_interval_behind = Spine::duration_string_to_minutes(aggregationIntervalStringBehind);
-    int agg_interval_ahead = Spine::duration_string_to_minutes(aggregationIntervalStringAhead);
-
-    if (agg_interval_behind < 0 || agg_interval_ahead < 0)
-      throw Fmi::Exception(BCP, "The 'interval' option must be positive!");
-
-    // set aggregation interval if it has not been set in parameter parser
-    for (const TS::ParameterAndFunctions& paramfuncs : poptions.parameterFunctions())
-    {
-      if (paramfuncs.functions.innerFunction.getAggregationIntervalBehind() ==
-          std::numeric_limits<unsigned int>::max())
-        const_cast<TS::DataFunction&>(paramfuncs.functions.innerFunction)
-            .setAggregationIntervalBehind(agg_interval_behind);
-      if (paramfuncs.functions.innerFunction.getAggregationIntervalAhead() ==
-          std::numeric_limits<unsigned int>::max())
-        const_cast<TS::DataFunction&>(paramfuncs.functions.innerFunction)
-            .setAggregationIntervalAhead(agg_interval_ahead);
-
-      if (paramfuncs.functions.outerFunction.getAggregationIntervalBehind() ==
-          std::numeric_limits<unsigned int>::max())
-        const_cast<TS::DataFunction&>(paramfuncs.functions.outerFunction)
-            .setAggregationIntervalBehind(agg_interval_behind);
-      if (paramfuncs.functions.outerFunction.getAggregationIntervalAhead() ==
-          std::numeric_limits<unsigned int>::max())
-        const_cast<TS::DataFunction&>(paramfuncs.functions.outerFunction)
-            .setAggregationIntervalAhead(agg_interval_ahead);
-    }
-
-    // store maximum aggregation intervals per parameter for later use
-    for (const TS::ParameterAndFunctions& paramfuncs : poptions.parameterFunctions())
-    {
-      std::string paramname(paramfuncs.parameter.name());
-
-      if (maxAggregationIntervals.find(paramname) == maxAggregationIntervals.end())
-        maxAggregationIntervals.insert(make_pair(paramname, AggregationInterval(0, 0)));
-
-      if (paramfuncs.functions.innerFunction.type() == TS::FunctionType::TimeFunction)
-      {
-        if (maxAggregationIntervals[paramname].behind <
-            paramfuncs.functions.innerFunction.getAggregationIntervalBehind())
-          maxAggregationIntervals[paramname].behind =
-              paramfuncs.functions.innerFunction.getAggregationIntervalBehind();
-        if (maxAggregationIntervals[paramname].ahead <
-            paramfuncs.functions.innerFunction.getAggregationIntervalAhead())
-          maxAggregationIntervals[paramname].ahead =
-              paramfuncs.functions.innerFunction.getAggregationIntervalAhead();
-      }
-      else if (paramfuncs.functions.outerFunction.type() == TS::FunctionType::TimeFunction)
-      {
-        if (maxAggregationIntervals[paramname].behind <
-            paramfuncs.functions.outerFunction.getAggregationIntervalBehind())
-          maxAggregationIntervals[paramname].behind =
-              paramfuncs.functions.outerFunction.getAggregationIntervalBehind();
-        if (maxAggregationIntervals[paramname].ahead <
-            paramfuncs.functions.outerFunction.getAggregationIntervalAhead())
-          maxAggregationIntervals[paramname].ahead =
-              paramfuncs.functions.outerFunction.getAggregationIntervalAhead();
-      }
-
-      if (maxAggregationIntervals[paramname].behind > 0 ||
-          maxAggregationIntervals[paramname].ahead > 0)
-        timeAggregationRequested = true;
-    }
   }
+  catch (...)
+  {
+    throw Fmi::Exception::Trace(BCP, "Operation failed!");
+  }
+}
+
+void Query::parse_attr(const Spine::HTTP::Request& theReq)
+{
+  try
+  {
+    string attr = Spine::optional_string(theReq.getParameter("attr"), "");
+    if (!attr.empty())
+    {
+      bool ind = true;
+      uint loopCount = 0;
+      while (ind)
+      {
+        loopCount++;
+        if (loopCount > 10)
+          throw Fmi::Exception(BCP, "The alias definitions seem to contain an eternal loop!");
+
+        ind = false;
+        std::string alias;
+        if (itsAliasFileCollectionPtr->replaceAlias(attr, alias))
+        {
+          attr = alias;
+          ind = true;
+        }
+      }
+
+      std::vector<std::string> partList;
+      splitString(attr, ';', partList);
+      for (const auto& part : partList)
+      {
+        std::vector<std::string> list;
+        splitString(part, ':', list);
+        if (list.size() == 2)
+        {
+          std::string name = list[0];
+          std::string value = list[1];
+
+          attributeList.addAttribute(name, value);
+        }
+      }
+    }
+    // ### Attribute list ( attr=name1:value1,name2:value2,name3:$(alias1) )
+    // attributeList.print(std::cout,0,0);
+  }
+  catch (...)
+  {
+    throw Fmi::Exception::Trace(BCP, "Operation failed!");
+  }
+}
+
+
+bool Query::parse_grib_loptions(const State& state)
+{
+  try
+  {
+	T::Attribute* v1 = attributeList.getAttributeByNameEnd("Grib1.IndicatorSection.EditionNumber");
+    T::Attribute* v2 = attributeList.getAttributeByNameEnd("Grib2.IndicatorSection.EditionNumber");
+
+    T::Attribute* lat = attributeList.getAttributeByNameEnd("LatitudeOfFirstGridPoint");
+    T::Attribute* lon = attributeList.getAttributeByNameEnd("LongitudeOfFirstGridPoint");
+
+    if (v1 != nullptr && lat != nullptr && lon != nullptr)
+    {
+      // Using coordinate that is inside the GRIB1 grid
+
+      double latitude = toDouble(lat->mValue.c_str()) / 1000;
+      double longitude = toDouble(lon->mValue.c_str()) / 1000;
+
+      std::string val = std::to_string(latitude) + "," + std::to_string(longitude);
+      Spine::HTTP::Request tmpReq;
+      tmpReq.addParameter("latlon", val);
+      loptions.reset(
+          new Engine::Geonames::LocationOptions(state.getGeoEngine().parseLocations(tmpReq)));
+	  return true;
+    }
+    else if (v2 != nullptr && lat != nullptr && lon != nullptr)
+    {
+      // Using coordinate that is inside the GRIB2 grid
+
+      double latitude = toDouble(lat->mValue.c_str()) / 1000000;
+      double longitude = toDouble(lon->mValue.c_str()) / 1000000;
+
+      std::string val = std::to_string(latitude) + "," + std::to_string(longitude);
+      Spine::HTTP::Request tmpReq;
+      tmpReq.addParameter("latlon", val);
+      loptions.reset(
+          new Engine::Geonames::LocationOptions(state.getGeoEngine().parseLocations(tmpReq)));
+	  return true;
+    }
+
+	return false;
+ }
+  catch (...)
+  {
+    throw Fmi::Exception::Trace(BCP, "Operation failed!");
+  }
+}
+
+void Query::parse_inkeyword_locations(const Spine::HTTP::Request& theReq, const State& state)
+{
+  try
+  {
+    auto searchName = theReq.getParameterList("inkeyword");
+    if (!searchName.empty())
+    {
+      for (const std::string& key : searchName)
+      {
+        Locus::QueryOptions opts;
+        opts.SetLanguage(language);
+        Spine::LocationList places = state.getGeoEngine().keywordSearch(opts, key);
+        if (places.empty())
+          throw Fmi::Exception(BCP, "No locations for keyword " + std::string(key) + " found");
+        inKeywordLocations.insert(inKeywordLocations.end(), places.begin(), places.end());
+      }
+    }
+ }
+  catch (...)
+  {
+    throw Fmi::Exception::Trace(BCP, "Operation failed!");
+  }
+}
+
+void Query::parse_origintime(const Spine::HTTP::Request& theReq)
+{
+  try
+  {
+    boost::optional<std::string> tmp = theReq.getParameter("origintime");
+    if (tmp)
+    {
+      if (*tmp == "latest" || *tmp == "newest")
+        origintime = boost::posix_time::ptime(boost::date_time::pos_infin);
+      else if (*tmp == "oldest")
+        origintime = boost::posix_time::ptime(boost::date_time::neg_infin);
+      else
+        origintime = Fmi::TimeParser::parse(*tmp);
+    }
+}
   catch (...)
   {
     throw Fmi::Exception::Trace(BCP, "Operation failed!");
