+++ timeseries/UtilityFunctions.cpp
@@ -4,6 +4,8 @@
 #include <boost/date_time/posix_time/posix_time.hpp>
 #include <timeseries/ParameterKeywords.h>
 #include <timeseries/ParameterTools.h>
+#include <engines/observation/ExternalAndMobileProducerId.h>
+#include <engines/observation/Keywords.h>
 
 namespace SmartMet
 {
@@ -13,166 +15,6 @@ namespace TimeSeries
 {
 namespace UtilityFunctions
 {
-// ----------------------------------------------------------------------
-/*!
- * \brief
- */
-// ----------------------------------------------------------------------
-
-void update_latest_timestep(Query& query, const TS::TimeSeriesVectorPtr& tsv)
-{
-  try
-  {
-    if (tsv->empty())
-      return;
-
-    // Sometimes some of the timeseries are empty. However, should we iterate
-    // backwards until we find a valid one instead of exiting??
-
-    if (tsv->back().empty())
-      return;
-
-    const auto& last_time = tsv->back().back().time;
-
-    if (query.toptions.startTimeUTC)
-      query.latestTimestep = last_time.utc_time();
-    else
-      query.latestTimestep = last_time.local_time();
-  }
-  catch (...)
-  {
-    throw Fmi::Exception::Trace(BCP, "Operation failed!");
-  }
-}
-
-// ----------------------------------------------------------------------
-/*!
- * \brief
- */
-// ----------------------------------------------------------------------
-
-void update_latest_timestep(Query& query, const TS::TimeSeries& ts)
-{
-  try
-  {
-    if (ts.empty())
-      return;
-
-    const auto& last_time = ts[ts.size() - 1].time;
-
-    if (query.toptions.startTimeUTC)
-      query.latestTimestep = last_time.utc_time();
-    else
-      query.latestTimestep = last_time.local_time();
-  }
-  catch (...)
-  {
-    throw Fmi::Exception::Trace(BCP, "Operation failed!");
-  }
-}
-
-// ----------------------------------------------------------------------
-/*!
- * \brief
- */
-// ----------------------------------------------------------------------
-
-void update_latest_timestep(Query& query, const TS::TimeSeriesGroup& tsg)
-{
-  try
-  {
-    // take first time series and last timestep thereof
-    TS::TimeSeries ts = tsg[0].timeseries;
-
-    if (ts.empty())
-      return;
-
-    // update the latest timestep, so that next query (is exists) knows from where to continue
-    update_latest_timestep(query, ts);
-  }
-  catch (...)
-  {
-    throw Fmi::Exception::Trace(BCP, "Operation failed!");
-  }
-}
-
-// ----------------------------------------------------------------------
-/*!
- * \brief
- */
-// ----------------------------------------------------------------------
-
-void store_data(TS::TimeSeriesVectorPtr aggregatedData, Query& query, TS::OutputData& outputData)
-{
-  try
-  {
-    if (aggregatedData->empty())
-      return;
-
-    // insert data to the end
-    std::vector<TS::TimeSeriesData>& odata = (--outputData.end())->second;
-    odata.emplace_back(TS::TimeSeriesData(aggregatedData));
-    update_latest_timestep(query, aggregatedData);
-  }
-  catch (...)
-  {
-    throw Fmi::Exception::Trace(BCP, "Operation failed!");
-  }
-}
-
-// ----------------------------------------------------------------------
-/*!
- * \brief
- */
-// ----------------------------------------------------------------------
-
-void store_data(std::vector<TS::TimeSeriesData>& aggregatedData,
-                Query& query,
-                TS::OutputData& outputData)
-{
-  try
-  {
-    if (aggregatedData.empty())
-      return;
-
-    TS::TimeSeriesData tsdata;
-    if (boost::get<TS::TimeSeriesPtr>(aggregatedData.data()))
-    {
-      TS::TimeSeriesPtr ts_first = *(boost::get<TS::TimeSeriesPtr>(aggregatedData.data()));
-      TS::TimeSeriesPtr ts_result(new TS::TimeSeries(ts_first->getLocalTimePool()));
-      // first merge timeseries of all levels of one parameter
-      for (const auto& data : aggregatedData)
-      {
-        TS::TimeSeriesPtr ts = *(boost::get<TS::TimeSeriesPtr>(&data));
-        ts_result->insert(ts_result->end(), ts->begin(), ts->end());
-      }
-      // update the latest timestep, so that next query (if exists) knows from where to continue
-      update_latest_timestep(query, *ts_result);
-      tsdata = ts_result;
-    }
-    else if (boost::get<TS::TimeSeriesGroupPtr>(aggregatedData.data()))
-    {
-      TS::TimeSeriesGroupPtr tsg_result(new TS::TimeSeriesGroup);
-      // first merge timeseries of all levels of one parameter
-      for (const auto& data : aggregatedData)
-      {
-        TS::TimeSeriesGroupPtr tsg = *(boost::get<TS::TimeSeriesGroupPtr>(&data));
-        tsg_result->insert(tsg_result->end(), tsg->begin(), tsg->end());
-      }
-      // update the latest timestep, so that next query (if exists) knows from where to continue
-      update_latest_timestep(query, *tsg_result);
-      tsdata = tsg_result;
-    }
-
-    // insert data to the end
-    std::vector<TS::TimeSeriesData>& odata = (--outputData.end())->second;
-    odata.push_back(tsdata);
-  }
-  catch (...)
-  {
-    throw Fmi::Exception::Trace(BCP, "Operation failed!");
-  }
-}
 
 bool is_special_parameter(const std::string& paramname)
 {
@@ -259,6 +101,59 @@ void get_special_parameter_values(const std::string& paramname,
     if (!timeseries.empty())
       result->push_back(
           TS::LonLatTimeSeries(Spine::LonLat(loc->longitude, loc->latitude), timeseries));
+  }
+}
+
+bool is_mobile_producer(const std::string& producer)
+{
+  try
+  {
+    return (producer == Engine::Observation::ROADCLOUD_PRODUCER ||
+            producer == Engine::Observation::TECONER_PRODUCER ||
+            producer == Engine::Observation::FMI_IOT_PRODUCER ||
+            producer == Engine::Observation::NETATMO_PRODUCER ||
+            producer == Engine::Observation::BK_HYDROMETA_PRODUCER);
+  }
+  catch (...)
+  {
+    throw Fmi::Exception(BCP, "Operation failed!", nullptr);
+  }
+}
+
+bool is_flash_producer(const std::string& producer)
+{
+  try
+  {
+    return (producer == FLASH_PRODUCER);
+  }
+  catch (...)
+  {
+    throw Fmi::Exception(BCP, "Operation failed!", nullptr);
+  }
+}
+
+bool is_icebuoy_or_copernicus_producer(const std::string& producer)
+{
+  try
+  {
+    return (producer == ICEBUOY_PRODUCER || producer == COPERNICUS_PRODUCER);
+  }
+  catch (...)
+  {
+    throw Fmi::Exception(BCP, "Operation failed!", nullptr);
+  }
+}
+
+bool is_flash_or_mobile_producer(const std::string& producer)
+{
+  try
+  {
+    return (is_flash_producer(producer) || is_mobile_producer(producer) ||
+            is_icebuoy_or_copernicus_producer(producer));
+  }
+  catch (...)
+  {
+    throw Fmi::Exception(BCP, "Operation failed!", nullptr);
   }
 }
